logger: Refactor logger to DB based implementation
The file based implementation for the log storage has a flaw -
the plugin folders are deleted when upgraded, so using them to store any data is problematic. Instead of using a file for storage we should use some other solution. One possible solution for this is having a separate table for our logs. This PR implements the database based solution.