php-compatibility: Improve compatibility with PHP versions
This PR fixes a mistakenly added named argument in function call. This was probably caused due to a not intentional double-click on the variable. To mitigate the issue I have set editor inlay hints to be visible only when holding ctrl + alt keys.

As the issue came out with PHPCompatibility rule that has not landed in any of the master versions. I scanned the codebase with the `development` branch of the package. I also fixed issues that came out regarding php 8.1 compatibility.