Drop advanced from logic
Dropping advanced form generation and display. This functionality is providing more complexity than the feature requires.

The feature was meant to allow customizing the form, but the average user who requires form customization doesn't know anything about customizing HTML. The tool is too advanced for non developers and doesn't provide any extra value for a developer. The functionality can be implemented using a built in functionality of custom HTML block.