Drop templater class and unify API queries
Custom templater class went against the "escape late" logic of WordPress. Templater was built so that the values provided to the templater should be escaped. Escaping should happen late and in the exact place the variable is echoed. Also refactoring the Smaily_Request class to improve API and expose available actions.