Render admin notices outside the form.
The admin notices are added below the .wrap class h1 and h2 tags. Currently this is the heading inside the form. This causes admin notices to appear inside the form element and therefore they pollute the POST request data. On certain cases the message content could break the form submissions - for example posting double nonce fields.

To avoid such cases we provide a container for admin messages outside the form element.