integrations/cf7: Multiform settings
The current implementation is wrong. Each form should be configurable individually allowing the user to use more than one form on the page. The current implementation makes it impossible to embed more than one form to your site as the configuration is shared between forms. This PR implements the functionality to configure forms individually by refactoring how the cf7 configuration options are saved to the database.