integrations/elementor: Reduce number of API calls
The elementor widget fronted makes API request to load autoresponders. This feature is only needed admin view where the user can select a autoresponder for the form. The selected autoresponder ID is saved and used when rendering the frontend widget. This PR improves the performance of the Elementor widget by allowing autoresponder fetching only in admin view.