rss: Add tax to RSS-feed items
RSS-feed items are mainly used to integrate with the Smaily built in template product importer. The main consumers are regular store buyers who need to pay taxes. Therefore the prices displayed in the newsletter should reflect the prices that the customer sees when opening the product link. This PR changes the RSS-item display logic to include prices with tax.