Fix: Improves patterns in .zipignore
The assets directory in .zipignore was ignored through the codebase. This caused right behavior in root folder where public assets were removed but also caused Elementor integration assets to be removed.

This PR forces all patterns to begin from the root of the workspace. This makes sure that folders are not omitted by accident.